Move SingleCodingChallenge into its own page
